`adjust::{pre,post}`に分割する
Refs: https://github.com/VOICEVOX/voicevox_core/pull/917#issuecomment-2570033651
