+++ crates/voicevox_core/src/core/adjust/post.rs
@@ -1,0 +1,10 @@
+pub(crate) fn ensure_minimum_phoneme_length(mut output: Vec<f32>) -> Vec<f32> {
+    const PHONEME_LENGTH_MINIMAL: f32 = 0.01;
+
+    for output_item in output.iter_mut() {
+        if *output_item < PHONEME_LENGTH_MINIMAL {
+            *output_item = PHONEME_LENGTH_MINIMAL;
+        }
+    }
+    output
+}
